Automatically report exit status in browser tests
This means that tests don't necessarily need to call REPORT_RESULT.

I'm going to be taking advantage of this in #12923.

Also, don't initialize EXIT_STATUS to zero, but leave it as undefined.
It seems useful to be able to tell the diference between a program that
exits with a zero status and one that has not in fact exited yet.