Fix JS pthreads proxying + WASM_BIGINT (#12935)
The proxying of JS methods has no type info. It used to just use a JS
Number (a double) for everything, which works, except with BigInts
that's no longer enough.

We have to serialize type info in this case. This does a simple thing of
prefixing each argument with a non-zero value if it is a bigint. This is
not very compact, but is maybe good enough for now.

Also add a HEAP64 global array which is useful here (and will be
with wasm64 too).

Fixes #12079