Remove EM_EXCLUSIVE_CACHE_ACCESS
While working on another change I noticed that this wasn't actually
working.  Specifically, it doesn't work when using mutli-processing to
compiler system libraries in parallel.

Here is why:

get_multiprocessing_pool() is called just once to setup the pool of
subprocess that emscripten uses.  sub-proccesses in this pool are
created just once and then re-used for various tasks.  Once a
sub-process has been created the parent process no longer has any
influsnce of its environment (os.environ).

Setting os.environ['EM_EXCLUSIVE_CACHE_ACCESS'] in the parent has no
effect on a pre-existing child process.  The sub-process pool will use
the value of EM_EXCLUSIVE_CACHE_ACCESS that was present when it was
first created.

This means that the value of `EM_EXCLUSIVE_CACHE_ACCESS` in the
sub-processes in the pool is fixed at pool creation time which is not
what we want.

Here is recipe for breaking it:

1. call get_multiprocessing_pool() without holding the lock.
   No need to use the pool here, just calling get_multiprocessing_pool()
   is enough to bring it into existence.
2. Aquire the lock lock with a view to building something using
   the pool.  This sets os.environ['EM_EXCLUSIVE_CACHE_ACCESS']
3. Build stuff with the pool.  None of the subprocess will
   have `EM_EXCLUSIVE_CACHE_ACCESS` and so will all block
   forever if they need the cache because the parent (which
   is waiting on them) holds the lock.

Possible solutions:

1. Destroy the pool each time the cache is locked or unlocked.
   This will cause it to be re-created.

2. Remove the EM_EXCLUSIVE_CACHE_ACCESS completely and avoid
   using the pool to perform operation that rely on the workers
   in the pool trying to get the lock.
   That is what this PR does.
   Perhaps set `EM_IN_WORKER` and assert if a worker in the pool ever
   tries to get the lock?