Run preRun callbacks on new pthreads
This is useful for my ongoing work to get dynamic linking working on
threads.  We use addOnPreRun to load side modules before any main module
code is run.  I imagine that other uses would probably also involve
setting up JS-side state before native code runs in which case I think
it makes sense to run these callback on each thread as it starts.

If it turns out that is not desirable we can design an alternative
mechanism for threads, but using this existing one makes the most sense
to me.

To implement this we delay the sending of the "loaded" reply from a new
thread until all the preRun callbacks have completed.