Followup on proxy_main cleanup
We had a single browser test set OFFSCREENCANVAS_SUPPORT=1 but
wasn't decorated with @requires_offscreen_canvas.

This test was relying on the bahaviour emscripten_proxy_main where
it would end up running main directly if pthread_create failed.
In thie case pthread_create fails because its tries to setup the
offscreen canvas.

This change avoids running that test when it offscree canvas is
not available and removes the fallback from emscripten_proxy_main.