Implement proxy handler for SPLIT_MODULE
For now, the proxy handler synchronously instantiates the deferred module, then
forwards the original call to the newly patched table slot. This PR extracts the
`instantiateSync` function out of `createWasm` so that the proxy handler can use
it as well.

Updates the existing SPLIT_MODULE test so that the profile will be empty,
causing all functions to be placed in the deferred module.