Reduce minmal pthread dependencies (#13048)
Explictly depend on the internal thread functions that we need, and
allow unused ones to be DCE'd.

The biggest win here was removing the forced dependency on vsnprintf
which was being pulled in via
`emscripten_sync_run_in_main_thread_xprintf_varargs`.  Interestingly
this function doesn't seem to have any callers at all but I assume
it has some use so leaving it in the code for now.