Don't change default CMake compiler / linker flags
It's a bit of a surprise that `Emscripten.cmake` overrides the
defaults for CMake's C/C++ compiler and linker flags. In particular
`RelWithDebInfo` by default doesn't include debugging information.
From the history of this file it wasn't clear why this decision was
made in the first place, but nowadays there doesn't seem to be a
good reason to keep doing it.