Avoid adding redundant `-c` and `-o` flags when compiling
This change move the handling of `-o` flags into the normal
parse_args loop.

Previously we had `find_output_arg` which was supposed to filter out
the `-o` flags but the output argument of this function were not
applied meaning that we were getting duplicate `-o` flags in many
cases.  These were harmless but odd when they appear in error messages
or the output `-v`.