Fix USE_OFFSET_CONVERTER + pthreads + O2
This change is required in order to fully enable asan+pthreads.

Move resetPrototype back to worker.js which is really the only place it
should be needed.  Calling resetPrototype must happen after the main
script is loaded (so that the WasmOffsetConverter class/function is
available).

Defining `resetPrototype` as a library function was problematic as it is
required by the worker script so would need to survive minification and
closure.  I think its simpler to define it right where its needed and
hooks into all code paths that signaled the that module is loaded.

It seems like we didn't have any tests for this so added one.