Fix name of file in `erase` method of harfbuzz and sdl2 ports
In practice this is not really a problem since this method is only
used when cleaning a port before building it, so that only side effect
would be that the old library continues to exist if the new build fails
for some reason.

The pattern of using a `get_lib_name` helper function is copied from the
regal port.