SPLIT_MODULE with dynamic linking (#13070)
Adds an INITIAL_MEMORY option that can be used to ensure that the instrumented
and split modules use the same table size. This is necessary because the table
is created in JS when using dynamic linking, so the table size is baked in to
the JS shared by the instrumented and split builds.

Also makes the dynamic linking infrastructure ignore all exports that start with
`%`. Users should pass --export-prefix='%' to wasm-split so that all of its
synthesized exports will make it through to the secondary module without
interference from the dynamic linker.