Avoid mangling + demangling in jsifier.js
We had a bug that only showed up with symbols that start with `$_`.
These happen to exist in `library_fetch.js` e.g:
`$__emscripten_fetch_xhr`.  In this case, when we mangle, we remove the
leading `$` giving `__emscripten_fetch_xhr`, and then if we demangle we
do that wrong thing and end up removing the `_` instead of re-adding the
  `$`.

This change avoids the round trip by always storing both the mangled
and demangled version of a given symbol.

See #12268