Fix strange/broken syntax in __resumeException__deps
Its hard for me to tell exactly what this was originally trying
to express, but right now it looks broken.

Prior to bcf60cea I think it might
have made more sense when the function was just one element of the
list, but then with bcf60cea it became a function with the list inside
it as expressions that do nothing.