Refactor gen_struct_info to allow separate compilation. NFC. (#13270)
As part of the pending musl update we can no longer (or at least it
becomes problematic) to compile internal musl headers alongside the
public headers.  To avoid this, this change updates gen_struct_info.py
to compile internal headers separately.

To facilitate this, this change adds support for compiling groups of
headers and merging the results into a single json output.

As part of debugging this I added support for compiling every header
file in isolation.  While this is slow, it uncovered a couple of
issues where structs and defines that were listed under the wrong
header and also a header that could not be compiled on its own
(SDL_pixels.h).