Acquire the cache lock before checking for sanity file
Without this check there is a race to create that sanity file when
emscripten if first used.   I noticed this strange behavior when I would
run the test suite just after clearig the cache.

On my machine the test suite would launch many emcc processes and each
of them would notice the sanity file as missing and run the checks.
Some of them would see an empty sanify file that one of the other
processes was in the middle of creating (before they flushed it), and
when this happens that process would try to clear the cache!  This
resulted in the cache being clearing by one process while it was being
populated by another.  Not good.

The fix is the hold the cache lock while we check for the sanity file
and while we run the sanity checks.

This problem has always existed but was recently exacerbated by the fact
that we populate the sysroot headers as soon as we start emcc.

Also remove superfluous ensure() calls.  Since we call this during
the constructor these days the cache can be assumed to always exist.