emcc: support merging static library into a single object (#7683)
Remove the optimization (again) for the special case of single object
input, which skips the linker and directly copies the input object file
to output.

This behavior interferes with linking (merging) all objects in a
library archive into a single object, done with the following command:

    cc -o lib.o -r -Wl,--whole-archive lib.a

The optimization was once useful when no clear distinction was made
between compile mode and link-only mode.  It was also required for
compatibility (see #9510 and #9571).

However, #9510 was addressed by #9511, and #9600 introduced the '-r'
flag that allows to explicitly indicate linking to an object, making the
optimization less useful.