Avoid writing even sanity.txt if FROZEN_CACHE is enabled
These improvements to the FROZE_CACHE setting allow for a
pre-populated read-only cache.  In thie senario the cache
is cannot ever be locked, and that we don't even write
the sanity file.

Users can still forece a sanity check with `emcc --check`
but no sanity file will ever be written.