Refactor Ports.fetch_port to avoid locking the cache where possible
This is in service of improving support for FROZEN_CACHE and is needed
as part of #13316.

The logic here mirrors the lock in cache.get() where we first attempt to
verify the contents of the cache without holding the lock and early out
in the common case the what we want is already here.