Add eslint for linting our build-time JS code (#13305)
Initially I chose a single file from the JS compiler and made it conform
to the style.

There is currently no plan to extend this to libray code or JS code that
runs on the client.  Partly for practical reasons, since that code all run
through a pre-processor, but also because we have very different JS
version requirements there.