Add HarfBuzz flag to SDL2_ttf rev 703ebc7c66fd. (#12868)
* Add HarfBuzz flag to SDL2_ttf rev 703ebc7c66fd.

`TTF_USE_HARFBUZ` flag adds complex script support
to SDL_ttf, hence the update to the latest revesion.

Fixes #12845.

* Add a new test for Arabic shaping with HarfBuzz.

This commit adds a test which uses SDL2_ttf for Arabic Shaping with
HarfBuzz. In order to do so, a font with Arabic Glyphs was necessary.
Therefore NotoNaskhArabic font was added.