Avoid setting noExitRuntime when using fibres/asyncify
`noExitRuntime` was getting set here and never getting unset meaning
that code that uses asyncify/fibres was never able to exit correctly.

Instead, we can use the new keepalive system to track when we there
is an outstanding unwound fibre/stack that could potentially be
rewound.  As long as there is potential to rewind a stack we don't
exit the runtime.

This system doesn't take into account multiple re-winding of the
same state/stack but AFAICT we don't even to that (yet).

Fixes: #14417