Build cached items directly to expected locations. NFC. (#13329)
When populating the cache with the creator functions that are passed to
`Cache.get` and `Cache.get_lib`, pass the name of the expected output
file to the creator function and assert that it creates it.  This avoid
unnessarily using a temporary file in most cases and saves some code
size/complexity.

Also add mpg123 to embuilder where it was missing.