[UBSan] Fix infinite recursion in type_info::operator==
Under UBSan, comparing type_infos triggers a lookup in UBSan's typeinfo cache.
When this cache is not warm, this causes a cache miss, and the code path that
handles that ends up calling __dynamic_cast. __dynamic_cast itself ends up
comparing type_infos and this recursion ends up exhausting the stack.

This is not a problem on native platforms because inlining and follow-on
optimizations remove the call to type_info::operator== under __dynamic_cast, so
UBSan never re-enters its type checking routine. In Emscripten, however, we
build libc++abi with -Oz, so these optimizations do not take place. The fix is
to add __attribute__((always_inline)) to type_info::operator==, which is enough
to get the recursion optimized out in -Oz.

Fixes #13330.