Use a specific lock file to avoid temp directory contention
Previously in debug mode we could try to avoid temp file contension by
taking and holding the cache lock for the duration of the process.

This change uses a specific lock file within the temp directory which
avoid needing a writable cache directory. Its also more logical.

Partial fix for #13369