[UBSan] Fix infinite recursion in type_info::operator== (#13367)
Under UBSan, comparing type_infos triggers a lookup in UBSan's typeinfo cache.
When this cache is not warm, this causes a cache miss, and the code path that
handles that ends up calling __dynamic_cast. __dynamic_cast itself ends up
comparing type_infos and this recursion ends up exhausting the stack.

```
RangeError: Maximum call stack size exceeded
    ...
    at std::type_info::operator==(std::type_info const&) const
    at is_equal(std::type_info const*, std::type_info const*, bool) 
    at __dynamic_cast (<anonymous>:wasm-function[40]:0x8fb)
    at __ubsan::checkDynamicType(void*, void*, unsigned long)
    at HandleDynamicTypeCacheMiss(__ubsan::DynamicTypeCacheMissData*,
        unsigned long, unsigned long, __ubsan::ReportOptions)
    at __ubsan_handle_dynamic_type_cache_miss
    at std::type_info::operator==(std::type_info const&) const
    ...
```

This is not a problem on native platforms because inlining and follow-on
optimizations remove the call to type_info::operator== under __dynamic_cast, so
UBSan never re-enters its type checking routine. In Emscripten, however, we
build libc++abi with -Oz, so these optimizations do not take place. The fix is
to add __attribute__((always_inline)) to type_info::operator==, which is enough
to get the recursion optimized out in -Oz.

Fixes #13330.
Fixes #13324.