Use cache directory when building system libraries (#13370)
This avoids the use of the temp directory when building
system libraries.  Instead we use a directory inside the cache.

At the point when we are building the system libray we are already
holding the cache lock.

This is part of a sequence of changes to improve FROZEN_CACHE
mode.