Save all temp files in EMCC_DEBUG mode
We have two classes of temp files that we use in emscripten

1. Temp files that live in emscripten temp dir and have non-unique
   names
2. Temp files that live direclyt in the temp dir (not in a subdirectory
   and must have unique names.

The cleanup mechanism we use for these two is separate.
- We keep (1) around when EMCC_DEBUG is set.
- We keep (2) around when EMCC_DEBUG_SAVE is set.

When EMCC_DEBUG is set we also use the emscripten temp dir (1) to store
the files of type (2), but we were still deleting them.

This change makes the saving of temp file consisten such that EMCC_DEBUG
acts as superset of EMCC_DEBUG_SAVE.

This has the nice property that repsonce files and other temporaries are
preserver in debug mode which makes subprocess commands more each to
reproduce.