Remove JS-only functions from deps_info.py (#13431)
I'm working on a test that will verify each of the entries in the
deps_info so that its impossible to becoming out-of-date or inaccruate.

This change is the first result of that.  These functions are JS-only
and not visible to native code.  Attempt to reference any of these
functions from native code result in link error because they are
prefixed with `$`.  This means they can never show up in the undefined
symbol list (the output of llvm-nm) of the valid program.