toolchain_profiler.py: Inject response files into reported commandline
Rather than appending the contents of the response file, inject it
at at the correct location, as the subprocess would do.

This also fixes and issue where `record_subprocess_spawn` was crashing
if a commandline was specified as a tuple rather than a list (since
tuples and lists connot be added).

See #13408.