Add smaller version of strtol with fewer dependencies
This also means that other functions that depend on strtol such
as inet_aton can be used without bringing in more code.

I measures the size of an otherwise empty program that only
references strtol and built with -Oz:

Before:
wasm: 827
js:   11510

After:
wasm: 2512
js:   11509

The minor change the test expectations actually brings the output
in line with that glibc produces so I think this might actually
be a bugfix too.