Always include native inet_aton (#13506)
This file was only being included under some circumstances
but its trivial wrapper around `inet_pton.c` which was always
being included in libc anyway.

This change helps with the musl upgrade as it means we can
build libsockets_proxy without the musl internal include
paths enabled.   The musl internal includes are problematic
when combined with non-musl code (such as the libsockets_proxy).