Restore line in newlocale.c that was lost in #4813
When musl was last upgraded this malloc line removed from newlocale.c,
perhaps by accident.

This restores the behaviour of musl which is to allow `newlocale` with
arbitrary names to succeed.  I verified that this is the case using musl
on my desktop.

The reason that test_locale_wrong was passing prior to #4813 is that
the musl older version of musl we were using prior to #4813 rejected
all locales except for "C" and "POSIX".

Upstream version that we are currently based on:
https://github.com/emscripten-core/musl/blob/v1.1.15/src/locale/newlocale.c#L44

Upstream version of musl that we were previously using:
https://github.com/emscripten-core/musl/blob/0b44a0315b47dd8eced9f3b7f31580cf14bbfc01/src/locale/newlocale.c#L5