Delay setting runtimeInitialized, even on pthreads
This means the getMemory in library_dylink.js (and potentially
other functions) work the same on the main the thread as on the
worker threads.   Without this getMemory falsely thinks the
runtime has been initialized and will call malloc rather than
doing static allocation when pre-loading shared libraries.

This solves part of the same issue that is solved by #13395.