Rewrite MinifyLocals from Uglify to Acorn. NFC (#13621)
Basically a simple rewrite from one AST to another, but also using slightly
more modern JS while doing so (Set, Map). Some parts differ though as
the ASTs are not compatible, so the algorithm is not 100% identical, but the
results should be.

Test changes are just minor whitespace, caused by the different whitespace
style that Uglify used to emit vs the new one.

Verified on wasm2js2, 3, s, z which we do not run on CI here.

After this we have just one remaining pass using Uglify, MinifyGlobals, after
which we can remove it entirely.