Rewrite MinifyGlobals from Uglify to Acorn (#13636)
This implements the same algorithm, with the same limitations as before,
namely it is not actually scope-aware, but wasm2js output is simple enough
that things work.

This does end up changing test outputs. It's very hard to avoid that as
the ASTs are different, and the order of iteration ends up different, and so
the minifed names are used in a different order. This should not matter.
Interestingly it improves code size, suggesting that the order used here
is more optimal, but that might be noise due to the small testcases.

This allows us to remove js-optimizer.js entirely and all the Uglify
dependency, but this PR leaves that for a subsequent NFC PR to
keep things smaller here.

Verified locally on wasm2js2, 3, s, z