Zero initialize static data section only in the main thread and add test for global static objects with dylink + pthreads. (#13395)
Along with #13603, this change fixes #13386

With #13603, the static data section for shared libraries is shared among different threads.
It also happened to fix global constructors being called in threads (additionally to the main thread).

This change fixes the remaining issue of the static data section being zero-initialized in each thread (additionally to the main thread). Fixing this remaining issue #13386 should be fixed.

This change also adds a test to ensure global static objects in shared libraries are handled correctly.