Add centralized handling for calling user callbacks (#13596)
This change also changes the way outstanding works keeps the runtime
alive.  We now use a counter rather than a single noExitRuntime variable
so that once outstanding works is complete the thread can be shutdown if
needed.

The specific need for this comes the use case of running fetch thread
and wanting to exit the thread once the fetch completes.  Without this
change `pthread_exit` from a fetch callback generated an `uncaught
unwind` message on the console.

In addition, it doesn't always make sense to call `pthread_exit`, (e.g.
in cases where you want C++ destructors on the stack to be called, or in
cases where C++ threads are being used and pthread_exit` does not exist,
or at least doesn't makes sense).  With this solution the thread will
exit automatically when returning from the fetch callback, no explicit
thread exit call needed.