Fix parentheses mix-up in batch files in Windows (#13698)
After #13498, `test_asyncify_onlylist_a` and `test_asyncify_onlylist_b`
started to failing. #13498 introduced `if ( ... ) else ( ... )` in
emcc/em++.bat, and it turns out the parentheses inside the command line
can be mixed with those `if`-`else`'s parentheses themselves. This does
not cause errors for every command line that contain parentheses, but
when there are both a comma and parentheses, this can trigger errors.

For example, `test_asyncify_onlylist_a`'s argument contains something
like this:
```
'ASYNCIFY_ONLY=["main",...,"foo(int,double)",...]'
```
Here the comma between `int` and `double` is mistaken as an item
separator within the list, and `)` after `double` is consequently
mistaken as ending `if`, because the whole body is within an `if`.

This PR fixes the problem by assigning those argument `%*` into a
variable `ARGS`, and within `if` and `else`, use it with a substitution
of `)` with the escaped end-paren `^)`.