Pass 1 explicitly to some set_setting() calls
`DISABLE_EXCEPTION_CATCHING` can take 0, 1, or 2. This is not an on-off
switches that only take 0 or 1, but when its argument is 1 it is
currently omitted. I think in this case it is more readable if we
actually pass the explicit argument. This is more of a drive-by fix and
does not intend to include all settings that can take any numeric
values.