Correctly fix windows parentheses mix-up (#13723)
This issue was attempted to fix in #13698 by escaping `)` but it turned
out to be an incomplete fix, because it fixed the breaking tests then
but broke other ones instead depending on their syntax. This PR tries to
fix it correctly by using `enabledelayedexpansion` feature of batch
files. We can enable delayed expansion of variables within a batch file
by
```
@setlocal enabledelayedexpansion
```
And if you use variables not by the usual syntax `%VAR%` but `!VAR!`, it
is expanded in a delayed manner, after all parentheses for `if`s and
`else`s are expanded.