Fix test_dasho_is_dir in Windows (#13724)
`test_dasho_is_dir` contains these three tests:
```
$ emcc tests/hello_world.c -o .                (a)
$ emcc tests/hello_world.c -o . --oformat=wasm (b)
$ emcc tests/hello_world.c -o . --oformat=html (c)
```

(a) errors out with
```
emcc: error: cannot write output file `.`: Is a directory
```
in all platforms. This message is printed by emcc.py.

(b) errors out with:
In Linux and Mac:
```
wasm-ld: error: cannot open output file .: Is a directory
```
In Windows:
```
wasm-ld: error: cannot open output file .: is a directory
```
Note that the `i`'s case in `is` is different. Note that this message is
not printed by emcc.py here; it is rather generated by wasm-ld, but it
doesn't seem to be printed from lld code. It looks it is printed from
the input/output system in the OSes, which explains the
uppercase/lowercase difference.

(c) errors out with:
In Linux and Mac:
```
emcc: error: cannot write output file: [Errno 21] Is a directory: '.'
```
In Windows:
```
emcc: error: cannot write output file: [Errno 13] Permission denied: '.'
```
So the `error: cannot write output file` part is printed by emcc.py, but
the "Is a directory" part is printed by the OS, and apparently Windows
prints a different message and error code.

So this PR fixes these problems by passing (b) and (c)'s
`assertContained` a list of two different messages so that either of
them can be found.