Fix settings.js doc for INLINING_LIMIT
This option was created as a numeric scale but currently what only
matters is if it is 0 or not, effectively making this a boolean switch.
This PR fixes the comment for `INLINING_LIMIT` in settings.js and
converts uses of `INLINING_LIMIT` options in tests that use values
greater than 1 to use 1.