Run emcmake and emconfigure via launcher scripts. NFC (#13768)
We only have couple of places where we were using `building.configure`
and `building.cmake` internally.  Its cleaner and simpler to run the via
a subprocess especially in the test suite so we get actual black box
testing.

As a followup we may be able move some of these functions into the
`emmake.py` and `emconfigure.py` scripts where it make more sense.