Update after Binaryen change to inline all single-use functions (#13744)
Always inlining single-use functions is beneficial, as in that case we can
delete the function after inlining, so the code is "moved". Binaryen's default
has changed to do that recently and this change updates our test expectations
for that.

Benefits:

As the updated test results here show, our code size tests shrink by 1-3%. The
metadce test changes show various functions that were inlined and then removed.

In addition, the emscripten benchmark suite shows a 3.5% code size reduction.
One big 30% dominates that; ignoring it using the median, the improvement is 2%,
and that remains the same when looking just at the large/real-world benchmarks.
Every single benchmark shrinks with this, and none regress.

Speed improves on a few benchmarks. That is mainly on the microbenchmarks,
though, so even while some of those speedups are large (one by 40%), I don't
think it's very realistic. The real-world codebases in the benchmark suite do
not change much in speed (but they do shrink by 2% as mentioned before).

Risks:

One reason we never did this is the risk of increasing compile times on VMs, due
to creating larger functions (and regalloc etc. can be nonlinear). I think that
risk is much lower today since we have great baseline compilers pretty much in
all wasm VMs. However, disabling the baseline compiler in v8 for example I can
see some startup regressions on the benchmark suite, like 10% on base64. So
there may be a downside here of slower times to reach the optimizing tier.
That's just in a few of the microbenchmarks, but it is a risk.

Even if we can delete the function after inlining we have no guarantee of a
total reduction in code size. That is the second reason we never did this. It is
hard to construct counterexamples manually, but the fuzzer easily does so when
instrumented for that. Investigating several of them, they seem like unlikely
scenarios. For example, in one case if we did not inline a function then the
target of the inlining would be identical to another function and duplicate
function elimination would end up helping more than inlining. My
impression is that these possible regressions are unlikely, especially since
every single benchmark in the benchmark suite shrinks as mentioned above.