Simplify library search process
When searching for libraries we were looking for both the `lib`
prefixed version and the non-`lib`-prefixed name.

While refactoring code in #13799 I noticed that we really don't need
to be handling both these cases.  We can assume that file on disc will
always have the lib prefix and the library names being processed here
will never have it.

This does remove some undocuments/untested (and I would have better
not supported) uses.

1. Linking with `-llibc` rather than `-lc` will no longer work.
2. Linking a library called `foo.a` (without a `lib` prefix) via `-lfoo`
   will not longer work.

Neither of these use cases is supported by gcc or clang and I don't
think it was ever intentional that we supported these.