Use cloud storage mirror of libjpeg archive
It turns out that the www.ijg.org refused requests from the urllib
library for some reason.

This issues was being masked by the fact that system_libs.py prefers to
use `requests` when it is installed.  This change undoes that.  This
feature was originally added in #8667 to deal with older python
installations but these days we supply our own python as part of emsdk
so we should be able to depend on a recent/correct version.  Having the
fallback here just makes testing harder.

Fixes: #13869