Remove setTempRet0, getTempRet0 from WASM_SYSTEM_EXPORTS (#13939)
This change allows more dynamic linking tests to build and run with
`MAIN_MODULE=2` rather than `MAIN_MODULE=2`.

Some background: `setTempRet0` and `getTempRet0` are a little odd
because they exist bot as runtime functions accessible from JS as
`setTempRet0` and `getTempRet0` but also as library functions accessible
as `_getTempRet0` and `_getTempRet0`.

The only JS symbols that can be linked with native code are JS library
symbols.  Runtime symbols like `setTempRet0` and `getTempRet0` are not
indended to be used to resolve undefined native symbols.  Undefined
native symbols will resolve to `_getTempRet0` and `_setTempRet0` in
jsifier (which unconditionally adds the `_` to JS library function
names).

More background: I believe `WASM_SYSTEM_EXPORTS` is designed to allow
certain symbols to be exported *from* wasm without mangling.  However,
these two symbols are always defined in JS and never defined in wasm so
I don't believe they ever belonged in this list.

The reason this fix is needed in order to enable more use of
`MAIN_MODULE=2` is that in `MAIN_MODULE=1` mode `getTempRet0` (the
runtime function is getting exported as part of `asmLibraryArgs`).
However in `MAIN_MODULE=2` mode we rely instead on the symbol existing
on the module object which it never does (without this change).  The
result is that we hit the default error:

```
  RuntimeError: abort('getTempRet0' was not exported. add it to EXPORTED_RUNTIME_METHODS (see the FAQ)) at Error
```

I plan on following up by removing the runtime versions of these
functions completely, but I wanted to land this change on its own
to make any issues easy to bisect.