Fix remaining issues reporting undefined symbols in MAIN_MODULE/RELOCATABLE/LINKABLE builds
Remove `gcc_personality_v0.c` from `libcompiler-rt`.  It contains a
bunch of references to libunwind function that we don't always include.
Apparently we use the one int `libcxxabi/src/cxa_personality.cpp` but we
don't depend on `gcc_personality_v0.c` at this point, so this is better
all around.

Also, allow programs built with `-sLINKABLE` to report undefined symbol
errors.  By default we don't report errors, but this can be overridden
on the command line).  The issue here was that `MAIN_MODULE=1` sets
`LINKABLE` which meant we were then not actually reporting  any
undefined symbols.