"fix" arcane closure madness
The problem here is that `e.target` may be (for whatever reason) the
`window` object. Closure may actually emit a global function named `id`,
which will be accessible through `window.id`. If both of these
conditions are met, `e.target.id ? e.target.id : ''` will evaluate to a
function value instead of a string, and we will end up calling `StringToUTF8`
on it, which will throw an exception.