Move shared library processing to link phase. NFC
This means this code doesn't run at all when the compiler
is running in `-c` (compile onlye) mode.

Also identify real wasm "dylib" files and don't filter them
in the same way that do for fake dylib files (basically
intermediate object files).  This is because they are always
safe to link in multiple times.