Do not error on deprecated functions in SIMD intrinsics test
The names of some intrinsic functions will soon change in upstream LLVM, but the
old names will be marked deprecated and preserved for some time. This PR should
smooth the transition by allowing use of the deprecated names without errors. A
future PR will update the deprecated names to the new names and remove this
exception.