Revert musl change to asctime.c. NFC
It looks like a patch from a more recent version of musl was applied
in #12043.

This change reverts that and removes the extra alias needed in in
library.js.  library.js now implements just a single `__asctime` which
is used by both `asctime.c` and `asctime_r.c`.

This is in the name doing more without musl matches and less in JS.