Remove unused implementedFunctions from jsifier metadata. NFC
This key is never assigned and we assert that its empty after running JS
compiler, so just remove it.

The purpose is served in amjs.js world in `jsifier.js` is taken care of
by checking WASM_EXPORTS in the block directly above the deleted code.

Also, remove unused isStubFunction and getRootIdent methods.