Remove LIBRARY_DEPS_TO_AUTOEXPORT setting
This settings seems to relate to asm.js.  It referes to symbols that
night be implemented in C, but the only symbol listsed by default
(`memcpy`) is *always* implemented in C these days.

The effect seems to be to export `memcpy` on the Module object in the
case that any JS function depends on memcpy.  I can't see why we want to
do that.