Deprecate `RUNTIME_LINKED_LIBS` setting. NFC (#14004)
Its better to simply list dynamic library dependencies directly on the
command line.

Handle the deprecated option by simply added the list of libraries to
the command line.  This means we no longer need to modify the dylink
section after linking since the linker will inject all the needed
dynamic libraries that it sees on the command line.