Split up monolithic run() method in emcc.py. NFC (#13998)
This is huge refactor that will cause all kind of merge conflicts for
folks who work on emcc.py but its really been a long time coming and is
going to main future maintenance much more manageable.

I basically split out each of the blocks that were being profiled into
separate functions, with a new EmccState object to hold any common
shared state.  A big advantage of this is that it becomes much more
clear what parts of the state are shared vs local.

The previous `run()` function was 1468 lines.  Now its only 140.