Improve undefined symbol reporting. NFC
We can be precise here when we report undefined symbols
rather then blindlyg searching the generated JS which
very hacky and inaccurate.

Before this change it was possible build with:

  `-s EXPORTED_FUNCTIONS=createWasm,_main`

And this would not generate an error because a function
called createWasm exists in the preamble.  However the
function would not be exported thus not honoring the
users request.  After this change a undefined symbol
error is correctly given.