Revert "Remove unused `hasExportedFunction` helper from  src/parseTools.js. NFC (#14009)"
This reverts commit da26539c7c048b33d1bf67517e8f54c85eb47dec.

Rather than remove `hasExportedFunction`, instead continue to use it in
our library code.

This change does still effect the semantics of the function but
hopefully in a way that makes it more accurate:  Previously we were
using `EXPORTED_FUNCTIONS` which is user-supplied list that can include
both native and JS symbols, and in some cases can contain symbols that
are not actually defined.  WASM_EXPORTS on the other had is an explicitly
list of native symbols derived directly from the compiled binary.