Fix args and return types of emmintrin.h
3 types of fixes here:
- use wasm_*_make functions instead of list initializing to avoid
depending on internal types
- cast args to v128_t before passing to Wasm SIMD instrinsics
- cast results of calling Wasm SIMD intrinsics to __m128/__m128i or
whatever the expected type is

I tested this by changing the typedef of __m128i to __i64x2 or v128_t
and compiling, with these changes, test_sse2 compiles and passes. I'm
not sure if we want to permanently typedef __m128i to v128_t to avoid
depending on internal types, I can do that in a follow-up.