Reinstate --cache flag in its original form
This change reverts #13954 and reinstates the `--cache`
flag much as it existed prior to #11094.  Essentially
this change reverts both of those changes.

This also allows us to re-enable the test_emar_response_file
test for windows.

We can look at landing more ubiquitous and robust repsonse
file processing that #13954 was aiming at as followup.