[Wasm2C] Support multiple libraries by using wasm2c prefixing (#14033)
wasm2c has a WASM_RT_ADD_PREFIX define that it looks for, and if
present it prefixes all exports etc. with it. That allows linking in multiple wasm2c
outputs into a single program. This PR uses the prefix in the additional support
code that we add, and includes a full test of linking in two prefixed libraries.

This basically works by compiling one library with prefix foo_ and another
with bar_, and then you can initialize one with foo_init(), call its exports
using foo_EXPORT_NAME() etc. etc.

This is mostly straightforward except for wasm_rt_register_func_type etc.
which wasm2c currently emits as extern - suggesting that the "runtime"
be shared among such linked libraries. That seems dangerous to me. For
now, get things to work by converting those to be static instead. That way
we have 100% modularity, but that is something we may want to discuss
at the wabt level.