Use `asmLibraryArg` when resolving symbols dynamically (#13949)
Use `asmLibraryArg` as the primary symbol map used in dynamic linking.
Previously we would use the `Module` object itself when looking up
symbols and when merging in new symbols.

After this change we continue to export new dynamic symbols on the
Module object but we might want to consider changing that in the future.

One of the issues that this fixes is that it allows JS library symbols
that were exported on `asmLibraryArg` but *not* exported on the Module
object (i.e. not included in DEFAULT_LIBRARY_FUNCS_TO_INCLUDE) to be
visible to the dynamic linker.  The test here requires that since it
needs to assign a GOT entry to a JS library function that is not
exported via DEFAULT_LIBRARY_FUNCS_TO_INCLUDE.

Fixes: #13786