circleci: macOS: don't try to brew install packages that are already installed
It seems that circleci recently started pre-installing python3 which
broke our CI with:

```
Error: python@3.9 3.9.0_1 is already installed
To upgrade to 3.9.4, run:
  brew upgrade python@3.9
```

This is kind of shame that brew can't just handle this case cleanly.  I
chose this approach over the `brew bundle` because it seemed a little
simple to me (someone who doesn't know brew or brew bundle).

See https://apple.stackexchange.com/questions/284379/with-homebrew-how-to-check-if-a-software-package-is-installed
See https://github.com/Homebrew/brew/issues/2491