Default to reported undefined symbol errors when building with MAIN_MODULE. (#14060)
For users that specify dynamic library dependencies on the command line
we can/should report undefined symbols at link time.

The only users that will, as a result of this change, need to add `-s
ERROR_ON_UNDEFINED_SYMBOLS=0` are those that specify or load libraries
at runtime using `loadWebAssemblyModule` or specifying
`dynamicLibraries` at runtime.  (Note that users of `dlopen` will not be
effected in this way).