Avoid closure minifying anything to "id" (#13965)
The problem is that e.target in our JS library code may be window object. Closure may
actually emit a global function named id, which will be accessible through window.id.
If both of these conditions are met, e.target.id ? e.target.id : '' will evaluate to a function
value instead of a string, and we will end up calling StringToUTF8 on it, which will throw
an exception.

To avoid that, tell closure to assume "id" is a name that is used by something already.