Add EXPORT_IF_DEFINED internal settings for soft exports
When a side module both imports and exports a symbol (this can happen
when it needs to address of functions or when a symbol is weakly
defined) we want to export those symbols from the main module if and
only if they are defined, but we don't want to generate errors when
they are not defined in the main module.

Normally this isn't a problem because we pass `--allow-undefined` but
after this change we can use `LLD_REPORT_UNDEFINED` in more dynamic
linking builds.